codegen: split main.go into several smaller files
